==> 2_Null_Documents.ipynb <==
# 1. Install, Imports, Settings

import pandas as pd
import numpy as np
from matplotlib import pyplot as plt

plt.rcParams["figure.figsize"] = [10, 6]
%config InlineBackend.figure_format = 'retina'

# 2. Load data

df = pd.read_csv("data/alertas.csv", sep="|")
print("Total Number of documents:", len(df))
print("Number of documents with no accesible text:", len(df[df['Text'].isnull()]))
print("Number of documents with accesible text:",  len(df[df['Text'].notnull()]))

df.head()

# 3. Number of Documents by Type, Subtype, Year 

at = df[df['Text'].notnull()].groupby(["Type", "Subtype", "Year"]).agg({"Filename":"count"})
at.columns = ["TxtAvailable"]

nat = df[df['Text'].isnull()].groupby(["Type", "Subtype", "Year"]).agg({"Filename":"count"})
nat.columns = ['NoTxtAvailable']

dfr = pd.concat([at, nat], axis=1)
dfr = dfr.fillna(0)
dfr["Total"] = dfr["TxtAvailable"] + dfr["NoTxtAvailable"]
dfr['%'] = round((dfr["NoTxtAvailable"] * 100) / (dfr["Total"]), 2)
dfr[dfr["%"] > 0]

# 7 Documentos con texto no disponible

df[df['Text'].isnull()]

list(df[df['Text'].isnull()]['Filename'])

# 7. Export HTML

!jupyter nbconvert --to html 2_Null_Documents.ipynb
