==> 7_NNA.ipynb <==
# 1. Install, Imports, Settings

#!pip install nltk
#!pip install rake_nltk
#!python -m spacy download es_core_news_md

import pandas as pd
import numpy as np
from matplotlib import pyplot as plt
import seaborn as sns
import spacy
from spacy.lang.es.examples import sentences 
nlp = spacy.load("es_core_news_md")
from collections import Counter

import re
from IPython.display import Markdown, display
import string, math
import nltk
#nltk.download('punkt')
from rake_nltk import Rake
from nltk.util import ngrams

plt.rcParams["figure.figsize"] = [10, 6]
%config InlineBackend.figure_format = 'retina'

punct = string.punctuation +'”“'

# 2. Loading data

df = pd.read_csv("data/alertas.csv", sep="|")

print("Total Number of documents:", len(df))
print("Number of documents with no accesible text (password protected):", len(df[df['Text'].isnull()]))
print("Number of documents with accesible text:",  len(df[df['Text'].notnull()]))

df.head()

# 3 Extracting Recommendations

rep = {"RECOMENDACIONES" : ["RECOMENDACIONE S", "RECOMENDACION ES", "RECOMENDACIO NES",
                            "RECOMENDACI ONES", "RECOMENDAC IONES", "RECOMENDA CIONES",
                            "RECOMEND ACIONES", "RECOMEN DACIONES", "RECOME NDACIONES",
                            "RECOM ENDACIONES", "RECO MENDACIONES", "REC OMENDACIONES",
                            "RE COMENDACIONES", "R ECOMENDACIONES",], 
        
       "recomendaciones":[ "recomendacione s", "recomendacion es", "recomendacio nes",
                            "recomendaci ones", "recomendac iones", "recomenda ciones",
                            "recomend aciones", "recomen daciones", "recome ndaciones",
                            "recom endaciones", "reco mendaciones", "rec omendaciones",
                            "re comendaciones", "r ecomendaciones",],
 
       "Recomendaciones":[  "Recomendacione s", "Recomendacion es", "Recomendacio nes",
                            "Recomendaci ones", "Recomendac iones", "Recomenda ciones",
                            "Recomend aciones", "Recomen daciones", "Recome ndaciones",
                            "Recom endaciones", "Reco mendaciones", "Rec omendaciones",
                            "Re comendaciones", "R ecomendaciones",],
        }


def replace_recomendaciones(text):
    text_res = text
    if text == text:
        for key in rep.keys():
            for value in rep[key]:
                if value in text:
                    text_res = text.replace(value, key)
    return text_res

df['Text'] = df['Text'].apply(replace_recomendaciones)

dff = df[df["Text"].notnull()]
i = 0
keywords = ["RECOMENDACIONES", "Recomendaciones", 
            "Al Gobernador de Bolívar, al Alcalde de Morales y a la Fuerza Pública",
            "En atención a lo anterior se recomienda:", 
            "el Sistema de Alertas Tempranas de la Defensoría del Pueblo formula las siguientes recomendaciones a las autoridades competentes:", 
            "las siguientes recomendaciones", 
            "las medidas que se enumeran a continuación", 
            "se recomienda lo siguiente", 
            "en particular se recomienda", 
            "1. A", 
            "adoptar las medidas que a continuación se enuncian", 
            "1- A", 
            "1) A", 
            "En dicha providencia, la Corte Suprema de Justicia ordenó:", 
            "Se recomienda:",
            "Comedidamente le solicito", 
            "1.AlaPolicía", 
            "Acontinuación sepresentan lasconclusiones ylasobservaciones alasacciones institucionales desarrolladas porlasautoridades delnivel nacional"
            
]

for k, t in dff.iterrows():
    c = i
    text = t['Text']
    for key in keywords:
        if key in text:
            #text[text.find(keyword):]
            i = i+1
            break
    if c == i and t['Subtype'] != "Notas" and t['Subtype'] != "Informe de Seguimiento":
        if ('recomendaciones' in text):
            print(t['Filename'])
            print(t['Text'])
            break
print('Number of documents with recommendations', i)
print('Number of documents without recommendations', len(dff) - i)

def split_recommendations(text, section):
    if text == text:
        for k in keywords:
            if k in text:
                return text[text.find(k):] if section else text[:text.find(k)]
    return text

df['Text_Original'] =  df['Text']
df['Recomendations'] = df['Text'].apply(split_recommendations, section=True)
df['Text'] = df['Text'].apply(split_recommendations, section=False)

# 4. Extracting NNA Sentences

children_keywords = ["NNAJ", "niños", "niñas", "adolescentes", "jóvenes", "NNA"]
agressions = ['homicidio', 'masacres', 'desplazamiento', 'desaparición forzada',] 
uso_recl = ['reclutamiento', 'expendio de droga', 'sicariato', "cobro de 'vacunas'", "vigilancia", 
            "acciones ejemplarizates", "utilización ilícita", "uso ilícito", "drogas ilícitas", "prácticas sexuales"]

# 4.1. Example of sentences that include use of NNAJ in the first 20 docs

text = ' '.join(df[df["Text"].notnull()]['Text'][20:50])
keyword = " NNAJ "
nlp.max_length = 2000000
document = nlp(text)
for sentence in document.sents:
    sentence = sentence.text
  
    if keyword.lower() in sentence.lower():
            
            #Use the regex library to replace linebreaks and to make the keyword bolded, again ignoring capitalization
            sentence = re.sub('\n', ' ', sentence)
            sentence = re.sub(f"{keyword}", f"**{keyword}**", sentence, flags=re.IGNORECASE)
            
            display(Markdown(sentence))

# 4.2 Extracting NNA sentences

def find_sentences_with_keyword(keywords, document):
    sents = []
    
    #Iterate through all the sentences in the document and pull out the text of each sentence
    for sentence in document.sents:
        sentence = sentence.text
        
        for keyword in keywords: 
            if keyword.lower() in sentence.lower():
            
                #Use the regex library to replace linebreaks and to make the keyword bolded, again ignoring capitalization
                sentence = re.sub('\n', ' ', sentence)
                sentence = re.sub(f"{keyword}", f"**{keyword}**", sentence, flags=re.IGNORECASE)
                sents.append(sentence)
                break           
    return sents

def extract_NNAJ_sentences(text):
    if text != text:
        return []
    document = nlp(text)
    return find_sentences_with_keyword(keywords=children_keywords, document=document)

df['NNAJ'] = df['Text'].apply(extract_NNAJ_sentences)

# 5. Saving partial text in csv

df['Recommendations'] = df['Recomendations']
df = df[['Filename', 'Text', 'Subtype', 'Type', 'Year', 'Path', 'Departamento',
       'Text_Original', 'Recommendations', 'NNAJ']]

df.columns

df.to_csv("data/alertas_NNA_sentences.csv", sep="|", index=False)

# 6. Export HTML

!jupyter nbconvert --to html 7_NNA.ipynb
