==> 8_NNAJ_NGrams.ipynb <==
# 1. Installs, Imports and Settings

#!pip install spacy==3.1.1 #restart runtime after this
#!python -m spacy download en_core_web_sm

import timer
import csv
import pandas as pd
pd.options.display.max_rows
pd.set_option('display.max_colwidth', None)

import string
import operator
from itertools import islice
from collections import Counter

#!pip install nltk
from nltk import ngrams
import nltk as nltk
from nltk.corpus import stopwords
#nltk.download('stopwords')
#nltk.download('punkt')
#nltk.download('averaged_perceptron_tagger')
#nltk.download('wordnet');

import spacy
nlp = spacy.load("es_core_news_md")

#!pip install pyenchant
import enchant

import time

#  2. Reading files 

df = pd.read_csv("data/alertas_NNA_sentences.csv", sep="|")

print("Size of Data:", len(df))
print()
print("Columns : ", list(df.columns))

# 3. Functions

exceptions = []
red_flags = []
green_flags = ['posdesmovilización', 'nnaj']
stop_words = stopwords.words('spanish') + list('abcdefghijklmñnopqrstuvwxyz') + red_flags
punct_signs = string.punctuation + '…¿•”“·*'

sim_dict = {'nnaj': ['niños', 'niñas', 'adolescentes', 'jovenes', 'jóvenes', 'niño', 'niña', 'j0ve', 'adolescente', 'joven'], 
            'reclutamiento' : ['eclutaniento'],
            'uso' : ['utilización'],
            'situación' : ['sit11 cién']
           }

d = enchant.Dict("es_CO")
def replace_similars(text, sim_dict = sim_dict):
    text = ' '.join(text)
    for key in sim_dict.keys():
        for v in sim_dict[key]:
            text = text.replace(v, key)
            
        text = text.replace(' '.join([key]*3), key)
        text = text.replace(' '.join([key]*2), key)

    return text.split()

def replace_accents(w):
    return w.lower().replace('á','a').replace('é','e').replace('í','i').replace('ó','o').replace('ú','u')

def check_spelling(text):
    new = []
    for token in text.split():
        if token in green_flags:
            new.append(token)
        elif d.check(token):
            new.append(token)
        elif (not d.check(token) and len(d.suggest(token)) > 0 and (replace_accents(token) == replace_accents(d.suggest(token)[0]))):
                new.append(d.suggest(token)[0].lower())
    return ' '.join(new)
            
def remove_exceptions(text):
    clean_text = text
    for e in exceptions:
        clean_text = clean_text.replace(e,' ')
    return clean_text

#clean stopwords and punctuation, lemmatization
def clean_text(text):   
    
    if type(text) == str:
        text = text.strip("']['").split("', '")
        
    if len(text) == 0:
        return ""
    
    tic = time.perf_counter()
    
    clean_text = ' '.join(text)
    clean_text = clean_text.translate(str.maketrans('', '', punct_signs))
    clean_text = check_spelling(clean_text)
    
    #clean_text = remove_exceptions (clean_text)
    nlp.max_length = len(clean_text) + 100
    doc = nlp(clean_text.lower())
    
    txt = []
    for token in doc:  
        if token.lemma_ not in stop_words and token.lemma:
            if token.pos_ in ['NOUN', 'ADJ', 'VERB']:
                txt.append(token.lemma_)
                    
    txt = replace_similars(txt)
    toc = time.perf_counter()  
    print(f"Performed in {toc - tic:0.4f} seconds")
    return txt

def get_list_nnaj_terms(text, n):
    if type(text) == str:
        text = text.strip("']['").split("', '")
        
    if len(text) == 0:
        return []

    ngramas = list(ngrams(text, n))
    
    return [' '.join (n) for n in ngramas]

clean_text(['violencia sexual es uno de los crimenes de guerra',
                     'drogas ilicitas y el mercado del dolar', 'anterior al dia de los muertos'])

## 4 Extract Ngrams

x = clean_text(df.loc[0]['NNAJ'])

list2_ = []
list1_ = df['NNAJ']

len(df)

for l in list1_:
    list2_.append (clean_text(l))

df['NNAJ_clean_text'] = list2_

df.to_csv("data/alertas_NNAJ_keywords", sep="|", index=False)

# 5. Extract NGRMAS

df = pd.read_csv("data/alertas_NNAJ_keywords", sep="|")
df.columns

df['NNAJ_keywords'] = df['NNAJ_clean_text'].apply(get_list_nnaj_terms,n=1)
df['NNAJ_keywords2'] = df['NNAJ_clean_text'].apply(get_list_nnaj_terms,n=2)
df['NNAJ_keywords3'] = df['NNAJ_clean_text'].apply(get_list_nnaj_terms,n=3)

df.to_csv("data/alertas_NNAJ_keywords.csv", sep="|", index=False)

df['NNAJ_keywords'].head(1)

df['NNAJ_keywords2'].head(1)

## 5. Exporting to html

!jupyter nbconvert --to html 8_NNAJ_NGrams.ipynb
