==> 14_Mapping_Actores_del_Conflicto_NNAJ.ipynb <==
# 1. Install, Imports, Settings

import pandas as pd
import numpy as np
from matplotlib import pyplot as plt
import seaborn as sns
import spacy
from spacy.lang.es.examples import sentences 
#!python -m spacy download es_core_news_md
nlp = spacy.load("es_core_news_md")
from collections import Counter
import string, math, json

plt.rcParams["figure.figsize"] = [10, 6]
%config InlineBackend.figure_format = 'retina'

punct = string.punctuation +'”“'

# 2. Loading data

grupos = {
    "Grupo Armado Organizado" : ["GAO"],
    "Grupo Armado Ilegal" : ["GAI"],
    "Grupo Armado Organizado Residual" : ["GAOR", "GAO-R", "grupo armado organizado - residual" ], 
    "Ejército de Liberación Nacional" : ["ELN"],
    "Autodefensas Gaitanistas de Colombia" : ["AGC"],
    "Fuerzas Armadas Revolucionarias - Ejército del Pueblo" : ["FARC", 
                                         "Fuerzas Armadas Revolucionarias de Colombia",
                                         "Fuerzas Armadas Revolucionarias Ejército del Pueblo",
                                         "FARCEP", "FARC-EP", "FARC EP"
                                        ],
    "Grupo delincuencial organizado" : ["GDO", ],
    "Jalisco Nueva Generación": [],
    "Bandas Criminales" : ["BACRIM", "Bandas Emergentes"],
    "Autodefensas Unidas de Colombia" : ["AUC", ],
    "Grupos posdesmovilización": [],
    "Ejército de Liberación Popular": ["EPL"],
    "Clan del Golfo" : ["AGC"],
}

children = ["NNAJ", "niños", "niñas", "adolescentes", "jóvenes", "NNA"]

narco_grupos  = ["Cartel del Amazonas",  
        "Cartel de Los Nevados",
        "Cartel de Bogotá",
        "Cartel de Buga",
        "Cartel de Medellín",
        "Oficina de Envigado",
        "Cartel de Cali",
        "Cartel del Caquetá",
        "Cartel del Norte del Valle",
        "Los Rastrojos",
        "Cartel de la Costa Atlántica",
        "Cartel de La Guajira",
        "Cartel del Cauca",
        "Los Caparrapos",
        "Los Paisas",
        "Águilas Negras",
        "Los Puntilleros",
        "Ejército Revolucionario Popular Antisubversivo de Colombia",
        "Los Pelusos",
        "GDO Los Pachenca",
        "Clan Gnecco Cerchar",
        "Clan Bustamante", 
        "Activo Primer Frente",
        "Segunda Marquetalia", 
        "Los Zetas", 
        "Cartel de Sinaloa", 
        "Oficina de Envigado", 
        "La Terraza",
        "Robledo",
        "Pachelly",
        "La Sierra", 
        "Caicedo", 
        "Los Triana",
        "Clan del Oriente", 
        "La Unión",
        "Clan del Norte",
        "Los Costeños", 
        "Los Pachencha", 
        "La Cordillera",
        "La Gran Alianza",
        "Banda Local",
        "Norte del Valle",
        "Los Chuckys",
        "Los Chacales",
        "La Constru",
        "Los Caqueteños"]

POI = {
    "Evaristo Porras": "Papa Doc",
    "Luis Agustín Caicedo Velandia" : "Don Lucho", 
    "Víctor Manuel Mejía Múnera": "Pablo Arauca",
    "Miguel Ángel Mejía Múnera": "El Mellizo",
    "Ramón Quintero":"",
    "Pablo Escobar" : "El patrón",
    "Gonzalo Rodríguez Gacha": "",
    "Diego Fernando Murillo": "",
    "Gilberto Rodríguez Orejuela": "El Ajedrecista",
    "Miguel Rodríguez Orejuela": "",
    "Leonidas Vargas": "El Rey de Caquetá",
    "Luis Enrique Calle Serna": "",
    "Wilber Varela": "",
    "Alberto Orlandez Gamboa": "", 
    "Hermágoras González": "",
    "Nicolás Rodríguez Bautista" : " Gabino ", 
    "Iván Márquez": "",
    "Gentil Duarte": "",
    "Jorge Gnecco Cerchar": "", 
    "Jesús María Aguirre" :"Chucho Mercancía",
    "Megateo": "",  
    "Vicente Castaño": "", 
    "Emiliano Alcides Osorio": " Caín ",
    "Edgar Bustamante" : "El Yuca",
    "Manuel Marulanda Vélez" :" Tirofijo ",     
    "Jorge Briceño Suarez Mono": " Jojoy ",
    "Dairo Antonio Úsuga David": "",
    "Salvatore Mancuso": "",
    "Rodrigo Tovar": "",
    "Carlos Castaño Gil": "",
    "Justo Pastor Perafán": "Don Pepe"
}

loc = pd.read_csv("resources/Departamentos_y_municipios_de_Colombia.csv")
loc.head()

dfnnaj = pd.read_csv("data/alertas_actors_NNAJ.csv", sep="|")
df = pd.read_csv("data/alertas_actors.csv", sep="|")
print("Total Number of documents:", len(df))
print("Total Number of documents NNAJ:", len(dfnnaj)) 

df.head()

dfnnaj.tail(5)

# 3. Cleaning Duty

arts = ['la', 'los', 'el', 'las', 'una', 'un']
def replace_accents(w):
    return w.replace('á','a').replace('é','e').replace('í','i').replace('ó','o').replace('ú','u')

arts = ['la', 'los', 'el', 'las', 'una', 'un']
def replace_articles(w):
    w = [x for x in w.split() if x not in arts]
    return ' '.join(w)

def clean_segment(s):
    return replace_articles(replace_accents(s.lower()))
    
def clean_actors(entities):
    entities = entities.strip("']['").split("', '")
   
    if len(entities) == 0:
        return []
    
    #grupos
    for key, value in grupos.items():
        for v in value:
            entities = [key if clean_segment(e) == clean_segment(v) else e for e in entities]
           
    #narco grupos
    for v in narco_grupos:
        entities = [v if clean_segment(e) == clean_segment(v) else e for e in entities]
    return entities

df["Actors"] = df["Actors"].apply(clean_actors)
dfnnaj["Actors_NNAJ"] = dfnnaj["Actors_NNAJ"].apply(clean_actors)

len(df), len(dfnnaj)

df.to_csv("data/alertas_actors.csv", sep="|")
df = df.set_index('Filename')
#remove duplicates
df = df[~df.index.duplicated(keep='first')]
df = df
df.head()

dfnnaj.to_csv("data/alertas_actors_NNAJ.csv", sep="|")
dfnnaj = dfnnaj.set_index('Filename')
dfnnaj.tail()

df = df.drop(['Unnamed: 0.1', 'Unnamed: 0'], axis = 1)
df.head()

#merging DFs
dftotal = pd.concat([dfnnaj, df.reindex(dfnnaj.index)], axis=1)
len(dftotal)

dftotal[dftotal['Actors_NNAJ'].isnull()]

dfnnaj = dfnnaj.reset_index()
dfnnaj.head()

len(dfnnaj)

dfnnaj.to_csv("data/alertas_actors_NNAJ.csv", sep="|", index=False)

# 4. Functions

df = pd.read_csv("data/alertas_actors_NNAJ.csv", sep="|")
df.head()

space_articles = []
red_flags = ['Telefax', 'procesoSubproceso Prevención y Protección Código','SAT', 'Alerta Temprana', 
             'Derecho Internacional Humanitario', 'SAT Código DPR', 
             'procesoSubproceso Prevención y Protección Código PPP02F10 ALERTA', 'SEGUIMIENTO', 
             'PARA', 'Frente', 'Nota de Seguimiento', 'PP', 'AT', 'LIBERTAD',  'Alertas Tempranas', 
             'P01 F07 Formato Nota de Seguimiento Versión 03 Vigente', 
             'Prevención y Protección Código PP', 'Bloque C', 'Comisión', 'POBLACION','Constitución Política', 
             'procesoSubproceso Prevención y Protección Código PP', 'Ejército', 'Sin', 'INFORME', 'MUSE', 
             'BIENES INDISPENSABLES PARA', 'Gobernación', 'A NO SER', 'cia', 'Alerta', 'RECOMENDACIONES A', 
            'NNA', 'Adolescentes', 'ADOLESCENTES', 'SAT INFORME DE RIESGO No']

arts = ['la', 'los', 'el', 'las']

def remove_flags(entities):
     return [e for e in entities if e not in red_flags]

def remove_articles(entities):
    new = []
    for i in range(0, len(entities) - 1):
            new.append(entities[i])
            for a in arts:
                if entities[i].lower().startswith(a + " "):
                    new[len(new)-1] = entities[i][len(a):]

    return new
                     
def fix_terms(entities):
    new = []
    for i in range(0, len(entities)):
        new.append(entities[i])
        # fix ELN
        if entities[i].startswith("ELN "):
                   new[len(new) - 1] =  "Ejército de Liberación Nacional"
        #fix Las FARC           
        elif entities[i].startswith("FARC") or entities[i].startswith("lasFARC") or entities[i].startswith("LasFARC"):
                   new[len(new) - 1] = "Fuerzas Armadas Revolucionarias - Ejército del Pueblo" 
                
        elif entities[i].startswith("Fuerzas Armadas Revolucionarias"):
                   new[len(new) - 1] = "Fuerzas Armadas Revolucionarias - Ejército del Pueblo" 
                
        elif entities[i].lower().startswith("urabe"):
                   new[len(new) - 1] = "Urabeños"     
                
        elif entities[i].startswith("UNP"):
                   new[len(new) - 1] = "Unidad Nacional de Protección"    
        
        elif entities[i].startswith("Acción Social"):
                   new[len(new) - 1] = "Agencia Presidencial para la Acción Social"
                
        elif entities[i].startswith("Al Instituto Colombiano de Bienestar Familiar") or entities[i].startswith("ICBF"):
                   new[len(new) - 1] = "Instituto Colombiano de Bienestar Familiar"
                
        elif entities[i].lower().startswith("armada"):
                   new[len(new) - 1] = "Armada Nacional"
                
        elif entities[i].lower().startswith("defensor"):
                   new[len(new) - 1] = "Defensoría del Pueblo"
                
    return new

def clean_list_actors(df, column):
    actors = df[column].sum()
    if type(actors) == str:
        actors = actors.strip("']['").split("', '")
    
    #remove articles
    actors = remove_articles(actors)
                     
    #fix some terms
    actors = fix_terms(actors)                               
                
    #remove red flags
    actors = remove_flags(actors)
    
    #remove blank spaces
    actors = [a.strip() for a in actors]
    
    return actors
    
def most_mentioned_actors(actors):           
    data = { 
             'actors': list(Counter(actors).keys()), 
             'frequency': list(Counter(actors).values())
           }

    counterdf = pd.DataFrame(data)
    counterdf = counterdf.groupby('actors').agg({'frequency' : 'sum'}).sort_values('frequency', ascending = False)
    return counterdf

# 4. Most Mentioned Actors - Top 20 in Documents that mention NNAJ

#### Los articulos se omiten por cuestiones de estandarizacion de los datos, pero por ejemplo Rastrojos se refiere a Los Rastrojos que es el nombre original, as'i pasa con Los Paisas, La Cordillera, La Constru, etc. 

dfa = most_mentioned_actors(clean_list_actors(df, 'Actors'))
dfa[:20]

# 4.1 Most Mentioned Actors - Top 20 in the Same Sentence that mention  NNAJ - More Context

dfb = most_mentioned_actors(clean_list_actors(df, 'Actors_NNAJ'))
dfb[:20]

df.head()

# 5. Statistical Distribution of Actors

dfa[dfa.frequency > 100].describe()

# 6. Top 10 actors in each period

df1 = df[df.Year < 2006]
df2 = df[(df.Year > 2005) & (df.Year < 2012)]
df3 = df[(df.Year > 2011) & (df.Year < 2017)]
df4 = df[df.Year > 2016]

# 6.1 2001 - 2005 (Conflicto, 2005 Justicia y Paz)

df1 = most_mentioned_actors(clean_list_actors(df1, 'Actors'))[:10]
df1

df1[df1.frequency > 100].describe()

# 6.2 2006- 2011 (Conflicto)

df2 = most_mentioned_actors(clean_list_actors(df2, 'Actors'))[:10]
df2

df2[df2.frequency > 100].describe()

# 6.3 2012 - 2016 (Negociaciones de Paz, Acuerdo de La Habana)

df3 = most_mentioned_actors(clean_list_actors(df3, 'Actors'))[:10]
df3

df3[df3.frequency > 100].describe()

# 6.4 2017 - 2022 (Post- Acuerdo)

df4 = most_mentioned_actors(clean_list_actors(df4, 'Actors'))[:10]
df4

df4[df4.frequency > 100].describe()

# 6. Export HTML

!jupyter nbconvert --to html 14_Mapping_Actores_del_Conflicto_NNAJ.ipynb
