==> 16_Actors_NNAJ_OverTime2.0.ipynb <==
# 1. Imports, Installs, Settings 

import seaborn 
import pandas as pd
import csv, numpy as np

# 2. Load Data

df = pd.read_csv("data/alertas_actors_NNAJ.csv", sep="|")
print(len(df))
df.head()

space_articles = []
red_flags = ['Telefax', 'procesoSubproceso Prevención y Protección Código','SAT', 'Alerta Temprana', 
             'Derecho Internacional Humanitario', 'SAT Código DPR', 
             'procesoSubproceso Prevención y Protección Código PPP02F10 ALERTA', 'SEGUIMIENTO', 
             'PARA', 'Frente', 'Nota de Seguimiento', 'PP', 'AT', 'LIBERTAD',  'Alertas Tempranas', 
             'P01 F07 Formato Nota de Seguimiento Versión 03 Vigente', 
             'Prevención y Protección Código PP', 'Bloque C', 'Comisión', 'POBLACION','Constitución Política', 
             'procesoSubproceso Prevención y Protección Código PP', 'Ejército', 'Sin', 'INFORME', 'MUSE', 
             'BIENES INDISPENSABLES PARA', 'Gobernación', 'A NO SER', 'cia', 'Alerta', 'RECOMENDACIONES A', 
            'NNA', 'Adolescentes', 'ADOLESCENTES', 'SAT INFORME DE RIESGO No']

arts = ['la', 'los', 'el', 'las']

def remove_flags(entities):
     return [e for e in entities if e not in red_flags]

def remove_articles(entities):
    new = []
    for i in range(0, len(entities) - 1):
            new.append(entities[i])
            for a in arts:
                if entities[i].lower().startswith(a + " "):
                    new[len(new)-1] = entities[i][len(a):]

    return new
                     
def fix_terms(entities):
    new = []
    for i in range(0, len(entities)):
        new.append(entities[i])
        # fix ELN
        if entities[i].startswith("ELN "):
                   new[len(new) - 1] =  "Ejército de Liberación Nacional"
        #fix Las FARC           
        elif entities[i].startswith("FARC") or entities[i].startswith("lasFARC") or entities[i].startswith("LasFARC"):
                   new[len(new) - 1] = "Fuerzas Armadas Revolucionarias - Ejército del Pueblo" 
                
        elif entities[i].startswith("Fuerzas Armadas Revolucionarias"):
                   new[len(new) - 1] = "Fuerzas Armadas Revolucionarias - Ejército del Pueblo" 
                
        elif entities[i].lower().startswith("urabe"):
                   new[len(new) - 1] = "Urabeños"     
                
        elif entities[i].startswith("UNP"):
                   new[len(new) - 1] = "Unidad Nacional de Protección"    
        
        elif entities[i].startswith("Acción Social"):
                   new[len(new) - 1] = "Agencia Presidencial para la Acción Social"
                
        elif entities[i].startswith("Al Instituto Colombiano de Bienestar Familiar") or entities[i].startswith("ICBF"):
                   new[len(new) - 1] = "Instituto Colombiano de Bienestar Familiar"
                
        elif entities[i].lower().startswith("armada"):
                   new[len(new) - 1] = "Armada Nacional"
                
        elif entities[i].lower().startswith("defensor"):
                   new[len(new) - 1] = "Defensoría del Pueblo"
                
    return new
    
def clean_list_actors(actors):
    if type(actors) == str:
        actors = actors.strip("']['").split("', '")
    
    #remove articles
    actors = remove_articles(actors)
                     
    #fix some terms
    actors = fix_terms(actors)                               
                
    #remove red flags
    actors = remove_flags(actors)
    
   # actors = find_missing(actors)
    
    #remove blank spaces
    actors = [a.strip().title() for a in actors]
    
    return actors

def most_mentioned_actors(actors):           
    data = { 
             'actors': list(Counter(actors).keys()), 
             'frequency': list(Counter(actors).values())
           }

    counterdf = pd.DataFrame(data)
    counterdf = counterdf.groupby('actors').agg({'frequency' : 'sum'}).sort_values('frequency', ascending = False)
    return counterdf

df['Actors_NNAJ'].head()

df['clean_NNAJ_actors'] = df['Actors_NNAJ'].apply(clean_list_actors)

f = open('actors.text', 'w', encoding = 'utf-8')
for x in sorted(list(set(df['clean_NNAJ_actors'].sum()))):
    f.write(x + '\n')

df.columns

dfa = df[['clean_NNAJ_actors', 'Filename', 'Year']]
dfa.head()

print(len(dfa))
dfa = dfa.explode('clean_NNAJ_actors')
len(dfa)

dfa.head()

dfa = dfa[dfa['clean_NNAJ_actors'].notnull()]
dfa.head()

non_armed_actors = [
 'Instituto Colombiano De Bienestar Familiar',
 'Fuerza Pública',
 'Policía Nacional',
 'Comisión Intersectorial Para La Prevención Del Reclutamiento',
 'Alcaldías',
 'Acción Comunal',
 'Corte Constitucional',
 'Fiscalía General De La Nación',
 'Policía',
 'Servicio Nacional De Aprendizaje',
 'Acción',
 'Agencia Presidencial Para La Acción Social',
 'Unidad Para La Atención',
 'Unidad Nacional De Protección',
 'Defensoría Del Pueblo',
 'Juntas',
 'Comisión Intersectorial',
 'Comisión Intersectorial Para La Prevención Del Reclutamiento Forzado',
 'Comisión Intersectorial De Prevención Del Reclutamiento',
 'Map',
 'Adolescentes Y Jóvenes',
 'Ejército De Liberación Popular',
 'Prevención Y Protección Código',
 'Grupos',
 'Gobernación De Bolívar',
 'Caicedo',
 'Departamento Administrativo Para La Prosperidad Social',
 'Al Servicio Nacional De Aprendizaje',
 'Asuntos Internacionales',
 'Consejo De Seguridad De Naciones Unidas',
 'Policía De Infancia',
 'Ejército Nacional',
 'Centros De Formación',
 'Jac',
 'Utilización Ilicita De',
 'Derechos Humanos',
 'Grupos Sociales Vulnerables Condición',
 'Unidad Administrativa Especial Para La Atención',
 'Uariv',
 'Sat Código Dprpc04 Versión Final Página',
 'Vicepresidencia De La República',
 'Comisión Intersectorial Para La Prevención Y Utilización De Niños',
 'No',
 'Spa',
 'Uso',
 'Instituto Colombiano Para El Bienestar Familiar',
 'Grupo Armado Organizado',
 'Reclutamiento',
 'Covid',
 'Gobernación De Arauca',
 'Plantilla Vigente',
 'Comisión Lntersectorial Para La',
 'Agencia Nacional De Tierras',
 'Unicef',
 'Procuraduría General De La Nación',
 'Comuna 13',
 'Utilización',
 'Seguimiento Vigente',
 'P01 F07',
 'Igualmente',
 'Jóvenes',
 'Unidad Móvil',
 'Coordinar',
 'Consejos Comunitarios',
 'Instituto Colombiano De Bienestar',
 'Nota De Seguimiento',
 'Adolescentes Desvinculados',
 'Acción Inmediata',
 'Reclutamientos',
 'Nnaj',
 'Implementar',
 'Oficina Del Alto Comisionado Para La Paz',
 'Adolescencia',
 'Directiva',
 'Instituto De Bienestar Familiar', 
'Al Instituto Colombiano',
 'De',
 'Corte',
 'Alcaldías Municipales',
 'Instituto Colombiano',
 'Fuerza Publica',
 'Srpa',
 'Niñas Adolescentes Y',
 'Consejo De Seguridad',
 'Armada Nacional',
 'Sede Central Calle 55 No',
 'Ejercito Nacional',
 'Fiscalía',
 'Generaciones Con Bienestar',
 'Sijin',
 'Prevención Del Reclutamiento',
 'Ruuvs',
 'Fuerzas Armadas',
 'Organismos De Seguridad Del Estado',
 'Fuerza',
 'Qu',
 'Pastoral Social',
 'Comisión Intersectorial De Alertas Tempranas',
 'Agencia Para La Reincorporación Y La Normalización',
 'Protección Social',
 'Procesosubproceso Prevención Y Protección Código Ppp01 F10',
 'Sat Código Dprpc04 Versión Última',
 'Carmen',
 'Acnur',
 'Cooperación Internacional',
 'Gobernación Del Meta',
 'Resolución A',
 'Red Nacional De Información',
 'Líderes',
 'Ciprat',
 'Facciones Disidentes',
 'Sat Código Dprfto07 Versión Final Página',
 'Al Ministerio De Defensa Nacional',
 'Al Instituto Colombiano De Bienestar Familia R',

 'Alcaldías Municipales De Bucaramanga',
 'Unidad De Atención',
 'Dirección Para La Acción Integral Contra Minas Antipersonal',
 'A',
 'Rni',
 'Secretaria Departamental De Educación',
 'Gobernación Del Huila',
 'Formato Nota De Seguimiento',
 'Organización Internacional Del Trabajo',
 'Infantería De Marina',
 'Ciprunna',
 'Estación',
 'Comisión Intersectorial De Prevención De Reclutamiento',
 'Secretario General',
 'Asamblea General De La Onu',
 'Al',
 'Unidad De Restitución De Tierras',
 'Policía Judicial',
 'Al Gobierno Nacional',
 'Conflicto',
 'Conflicto Armado',
 'Educación',
 'Nbi',
 'Unión Peneya',
 'Departamento Para La Prosperidad Social',
 'Miembros',
 'Policía Departamental De Arauca',
 'Secretaría Técnica', 
 'Icb F', 
'Riesgo Vigente', 
'Gitó Dokabú',
'Prevención Y Protección', 
'Robledo', 
 'Consejo De Seguridad De Las Naciones Unidas',
 'Gobernación De Antioquia',
 'Constitución',
 'Institución Educativa',
 'Subcomité De Prevención Protección',
 'Ministerio De Defensa',
 'Alcaldías De El Espinal',
 'Articulación De La Oferta',
 'Ddhh',
 'Comité Interinstitucional De Alertas Tempranas',
 'Eai',
 'Registraduría General De La Nación',
 'Nota De Seguimiento N°', 
 'Al Ministerio De Agricultura',
 'Ministro De Defensa Nacional',
 'No2127',
 'Protección De La Niñez',
 'Masac',
 'L As Farc',
 'Odulio',
 'Oficina En',
 'Organización',
 'Osc',
 'Pr Oferidas',
 'Pri',
 'Junta',
 '5 A',
 'Reflejo Guásima',
 'Re9Ional',
 'Social Fe',
 'Unión Baquiaza',
 'Tres Jóvenes',
 'Tra Vés',
 'Tra',
 'Subcomité De Juventud',
 'Sergio Leder',
 'Red De Solidaridad Social',
 'Secretaria De Salud Del Valle',
 'Sama A Na',
 'Said Galvis',
 'Sa Email',
 'Revivir Al Futu Ro',
 'Inspección De La Esmeralda',
 'Iu',
 'Fuerza Aérea',
 'Garci A',
 'Comité De Cafeteros',
 'Administración Municipal',
 'Alto Comisionado De Las Naciones Unidas Para Los Derechos Humanos',
 'As Farc',
 'Atropetos Ll Ccntra',
 'Awà',
 'Cai’S',
 'Cerro Azul',
 'Cmtrol',
 'Colombia Email',
 'Comité Lnterinstitucional De Alertas Tempranas',
 'Fundación',
 'Consejos Comunitarios Afrodescendientes',
 'Corporación Huellas',
 'Curundó Carmen Del Surama',
 'Departamento De Policia De Cundinamarea Descripción Del Riesgo Se',
 'Departamento De Policía De Cundinamarca',
 'Dijin',
 'Epinayú',
 'Forzoso',
 'A La Propiedad',
 'Fuerza De Tarea',
 'Visión Mundial', 
'Comité Interins',
'Programa',
'Al Ministerio De Educación', 
'República', 
'Comisarías De Familia', 
'Naciones Unidas',
'Comité Interinstitucional Para La Articulación De La Oferta', 
'Unidad Nacional De Atención', 
    'Desplazamiento Forzado', 
'Justicia', 
    'Gobernación Del Departamento', 
    'Departamento Administrativo De Seguridad', 
    'Comité Departamental De Justicia Transicional',
'Notas De Seguimiento',
'Comuna Villa Santana',
'Gobernación De Caquetá',
    'Sierra',
    'Unión',  
    'Norte Del Valle',
    'Unión Patriótica', 
]

dfa = dfa[~dfa.clean_NNAJ_actors.isin(non_armed_actors)]
dfa.clean_NNAJ_actors = dfa.clean_NNAJ_actors.apply(lambda s: 'Auc' if s == 'Au C' else s)

dfa.clean_NNAJ_actors = dfa.clean_NNAJ_actors.apply(lambda s: 'Autodefensas Unidas De Colombia' if s == 'Autodefensas Unidas De Colom' else s)

dfa.head()

# Adding classificación de María Paula

categories  = {'Guerrillas' : ['Fuerzas Armadas Revolucionarias - Ejército Del Pueblo', 'Ejército De Liberación Nacional', 
                              'Frente Cacique Pipintá', 'Frente Capital', 'Frente 36 Delas Farcylos', 
                              'Frente 6', 
                              ], 
               
               'Paramilitares' : [ 'Autodefensas Unidas De Colombia', 'Autodefensas', 'Bloque Centauros', 'Bloque Calima', 
                                 'Frente Capital Bloque Centauros', 'Frente Capital De Las Auc', 'Autodefensas Del Bloque Central Santander', 
                                  'Bloque Centauros De Las Auc', 'Bloque Capital', 'Bloque Centauros', 
                                  
                                 ], 
               
               'Disidensias FARC' : ['Frente Primero','Frente Primero De Las Farc',
                                     'Facciones Disidentes De Las Ex Farc',  'Frente De Guerra Occidental',],
               
               'Nuevos Paramilitares (Narco-Para)' : ['Rastrojos', 'Águilas Negras', 'Autodefensas Gaitanistas De Colombia', 
                                                     'Urabeños', 'Clan Del Golfo', 'Bandas Criminales', 
                                                      'Autodefensas Gaitanistas', 'Gaitanistas', 'Autodefensas', 'Bvpa', 'Clan', 
                                                     'Gaitanistas De Colombia', 'Grupos Posdesmovilización', ], 
               
               'Pequeños Grupos Ilegales' : ['Constru', 'Paisas', 
                                             'Grupos Armados De Crimen Organizado', 'Grupos Delictivos Organizados', 
                                             'Jóvenes Por Grupos Armados Organizados', 'Bandas Criminales', 'Cordillera', 'Terraza'
                                             'Grupos Armados Organizados', 
                                             'Grupos Armados Irregulares', 
                                             'Oficina De Envigado', 'Terraza'
                                             
                                            ]
               
              }

def get_cat(actor):
    for k in categories.keys():
        if actor in categories[k]:
            return k
    return np.nan

dfa['categoria'] = dfa['clean_NNAJ_actors'].apply(get_cat)

dfa

terms = ['Autodefensas Gaitanistas', 'Gaitanistas', 'Gaitanistas De Colombia']
dfa['clean_NNAJ_actors'] = dfa.clean_NNAJ_actors.apply(lambda s: 'Autodefensas Gaitanistas De Colombia' if s in terms else s)

# Saving csv file 

dfa.to_csv('data/clean_NNAJ_actors.csv', sep='|', index = False)

dfa.head()
